move boxbody to own mod
